Fix types

+++ c/include/libsbp/v4/bootload/MSG_BOOTLOADER_HANDSHAKE_DEP_A.h
@@ -92,7 +92,7 @@ int sbp_msg_bootloader_handshake_dep_a_handshake_strcmp(
  * @return Size of sbp_msg_bootloader_handshake_dep_a_t::handshake in wire
  * representation
  */
-uint8_t sbp_msg_bootloader_handshake_dep_a_handshake_encoded_len(
+size_t sbp_msg_bootloader_handshake_dep_a_handshake_encoded_len(
     const sbp_msg_bootloader_handshake_dep_a_t *msg);
 
 /**
@@ -105,7 +105,7 @@ uint8_t sbp_msg_bootloader_handshake_dep_a_handshake_encoded_len(
  * @param msg sbp_msg_bootloader_handshake_dep_a_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_bootloader_handshake_dep_a_handshake_space_remaining(
+size_t sbp_msg_bootloader_handshake_dep_a_handshake_space_remaining(
     const sbp_msg_bootloader_handshake_dep_a_t *msg);
 /**
  * Set sbp_msg_bootloader_handshake_dep_a_t::handshake
@@ -207,8 +207,8 @@ const char *sbp_msg_bootloader_handshake_dep_a_handshake_get(
  * @param msg sbp_msg_bootloader_handshake_dep_a_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_bootloader_handshake_dep_a_handshake_section_strlen(
-    const sbp_msg_bootloader_handshake_dep_a_t *msg, uint8_t section);
+size_t sbp_msg_bootloader_handshake_dep_a_handshake_section_strlen(
+    const sbp_msg_bootloader_handshake_dep_a_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_bootloader_handshake_dep_a_t
