+++ c/include/libsbp/v4/file_io/MSG_FILEIO_READ_REQ.h
@@ -109,7 +109,7 @@ int sbp_msg_fileio_read_req_filename_strcmp(const sbp_msg_fileio_read_req_t *a,
  * @param msg sbp_msg_fileio_read_req_t instance
  * @return Size of sbp_msg_fileio_read_req_t::filename in wire representation
  */
-uint8_t sbp_msg_fileio_read_req_filename_encoded_len(
+size_t sbp_msg_fileio_read_req_filename_encoded_len(
     const sbp_msg_fileio_read_req_t *msg);
 
 /**
@@ -122,7 +122,7 @@ uint8_t sbp_msg_fileio_read_req_filename_encoded_len(
  * @param msg sbp_msg_fileio_read_req_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_fileio_read_req_filename_space_remaining(
+size_t sbp_msg_fileio_read_req_filename_space_remaining(
     const sbp_msg_fileio_read_req_t *msg);
 /**
  * Set sbp_msg_fileio_read_req_t::filename
@@ -219,8 +219,8 @@ const char *sbp_msg_fileio_read_req_filename_get(
  * @param msg sbp_msg_fileio_read_req_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_fileio_read_req_filename_section_strlen(
-    const sbp_msg_fileio_read_req_t *msg, uint8_t section);
+size_t sbp_msg_fileio_read_req_filename_section_strlen(
+    const sbp_msg_fileio_read_req_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_fileio_read_req_t
