+++ c/include/libsbp/v4/file_io/MSG_FILEIO_REMOVE.h
@@ -90,7 +90,7 @@ int sbp_msg_fileio_remove_filename_strcmp(const sbp_msg_fileio_remove_t *a,
  * @param msg sbp_msg_fileio_remove_t instance
  * @return Size of sbp_msg_fileio_remove_t::filename in wire representation
  */
-uint8_t sbp_msg_fileio_remove_filename_encoded_len(
+size_t sbp_msg_fileio_remove_filename_encoded_len(
     const sbp_msg_fileio_remove_t *msg);
 
 /**
@@ -103,7 +103,7 @@ uint8_t sbp_msg_fileio_remove_filename_encoded_len(
  * @param msg sbp_msg_fileio_remove_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_fileio_remove_filename_space_remaining(
+size_t sbp_msg_fileio_remove_filename_space_remaining(
     const sbp_msg_fileio_remove_t *msg);
 /**
  * Set sbp_msg_fileio_remove_t::filename
@@ -201,8 +201,8 @@ const char *sbp_msg_fileio_remove_filename_get(
  * @param msg sbp_msg_fileio_remove_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_fileio_remove_filename_section_strlen(
-    const sbp_msg_fileio_remove_t *msg, uint8_t section);
+size_t sbp_msg_fileio_remove_filename_section_strlen(
+    const sbp_msg_fileio_remove_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_fileio_remove_t
