+++ c/include/libsbp/v4/logging/MSG_LOG.h
@@ -93,7 +93,7 @@ int sbp_msg_log_text_strcmp(const sbp_msg_log_t *a, const sbp_msg_log_t *b);
  * @param msg sbp_msg_log_t instance
  * @return Size of sbp_msg_log_t::text in wire representation
  */
-uint8_t sbp_msg_log_text_encoded_len(const sbp_msg_log_t *msg);
+size_t sbp_msg_log_text_encoded_len(const sbp_msg_log_t *msg);
 
 /**
  * Query sbp_msg_log_t::text for remaining space
@@ -105,7 +105,7 @@ uint8_t sbp_msg_log_text_encoded_len(const sbp_msg_log_t *msg);
  * @param msg sbp_msg_log_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_log_text_space_remaining(const sbp_msg_log_t *msg);
+size_t sbp_msg_log_text_space_remaining(const sbp_msg_log_t *msg);
 /**
  * Set sbp_msg_log_t::text
  *
@@ -196,8 +196,8 @@ const char *sbp_msg_log_text_get(const sbp_msg_log_t *msg);
  * @param msg sbp_msg_log_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_log_text_section_strlen(const sbp_msg_log_t *msg,
-                                        uint8_t section);
+size_t sbp_msg_log_text_section_strlen(const sbp_msg_log_t *msg,
+                                       size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_log_t
