+++ c/include/libsbp/v4/piksi/MSG_COMMAND_REQ.h
@@ -94,7 +94,7 @@ int sbp_msg_command_req_command_strcmp(const sbp_msg_command_req_t *a,
  * @param msg sbp_msg_command_req_t instance
  * @return Size of sbp_msg_command_req_t::command in wire representation
  */
-uint8_t sbp_msg_command_req_command_encoded_len(
+size_t sbp_msg_command_req_command_encoded_len(
     const sbp_msg_command_req_t *msg);
 
 /**
@@ -107,7 +107,7 @@ uint8_t sbp_msg_command_req_command_encoded_len(
  * @param msg sbp_msg_command_req_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_command_req_command_space_remaining(
+size_t sbp_msg_command_req_command_space_remaining(
     const sbp_msg_command_req_t *msg);
 /**
  * Set sbp_msg_command_req_t::command
@@ -204,8 +204,8 @@ const char *sbp_msg_command_req_command_get(const sbp_msg_command_req_t *msg);
  * @param msg sbp_msg_command_req_t instance
  * @return Length of section
  */
-uint8_t sbp_msg_command_req_command_section_strlen(
-    const sbp_msg_command_req_t *msg, uint8_t section);
+size_t sbp_msg_command_req_command_section_strlen(
+    const sbp_msg_command_req_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_command_req_t
