+++ c/include/libsbp/v4/settings/MSG_SETTINGS_READ_RESP.h
@@ -95,7 +95,7 @@ int sbp_msg_settings_read_resp_setting_strcmp(
  * @param msg sbp_msg_settings_read_resp_t instance
  * @return Size of sbp_msg_settings_read_resp_t::setting in wire representation
  */
-uint8_t sbp_msg_settings_read_resp_setting_encoded_len(
+size_t sbp_msg_settings_read_resp_setting_encoded_len(
     const sbp_msg_settings_read_resp_t *msg);
 
 /**
@@ -108,7 +108,7 @@ uint8_t sbp_msg_settings_read_resp_setting_encoded_len(
  * @param msg sbp_msg_settings_read_resp_t instance
  * @return Maximum number of bytes that can be appended to the existing string
  */
-uint8_t sbp_msg_settings_read_resp_setting_space_remaining(
+size_t sbp_msg_settings_read_resp_setting_space_remaining(
     const sbp_msg_settings_read_resp_t *msg);
 /**
  * Return the number of sections in sbp_msg_settings_read_resp_t::setting
@@ -116,7 +116,7 @@ uint8_t sbp_msg_settings_read_resp_setting_space_remaining(
  * @param msg sbp_msg_settings_read_resp_t instance
  * @return Number of sections in string
  */
-uint8_t sbp_msg_settings_read_resp_setting_count_sections(
+size_t sbp_msg_settings_read_resp_setting_count_sections(
     const sbp_msg_settings_read_resp_t *msg);
 
 /**
@@ -241,7 +241,7 @@ bool sbp_msg_settings_read_resp_setting_append_vprintf(
  * @return Pointer to C string, NULL on error
  */
 const char *sbp_msg_settings_read_resp_setting_get_section(
-    const sbp_msg_settings_read_resp_t *msg, uint8_t section);
+    const sbp_msg_settings_read_resp_t *msg, size_t section);
 
 /**
  * Obtain the length of a section in sbp_msg_settings_read_resp_t::setting
@@ -254,8 +254,8 @@ const char *sbp_msg_settings_read_resp_setting_get_section(
  * @param section Section number
  * @return Length of section
  */
-uint8_t sbp_msg_settings_read_resp_setting_section_strlen(
-    const sbp_msg_settings_read_resp_t *msg, uint8_t section);
+size_t sbp_msg_settings_read_resp_setting_section_strlen(
+    const sbp_msg_settings_read_resp_t *msg, size_t section);
 
 /**
  * Get encoded size of an instance of sbp_msg_settings_read_resp_t
